(#5)[feat]: Added pause menu with buttons
I've added an inventory, continue game, and settings button

Unfortunately I have not been able to have the menu be invisible in the
editor.
Setting the Display to none or Visibility to Hidden in the uxml by default, leaves me unable to set them back to any other value in c# code.
So the display style is set to none in the Start() function of my UI
script.