(#5 && #4)[feat]: Added Intel for enemy pickups and sneaking
For the inventory functionality I need items to gather and test.
So it's both a commit for the menus and enemy pickups
I used the next part of the guide I followed previously to figure out how to do ScriptableObjects: https://www.whatupgames.com/blog/code-the-grid-based-inventory-system
One thing to note if any developers are reading this, a ScriptableObject derived class can't be abstract, I guess that makes sense if assets use some type of instance of your ScriptableObject derived class (You can't make instances of an abstract class.