feature: added v2 player and camera + added compat scene/mode for v1
v2 camera currently follows and looks at the player with all the default cinemachine Free Look Camera settings
also with how the v2 movement script is coded, jumping throws an
exception

v1 compatibility scene and script changes are pretty scuffed but at least it works which is important right now

I've opted to use the DEBUG preprocessor directive to avoid any
compatibility code to ever run in a release build, but sometimes I did
have to add a v1 script field and check if it's null.

What's most important with compatiblity is that I should be able to play
the game using the V1 player on the v1 scene while also playing the game
on the current scene which contains the new v2 stuff without losing any
performance by doing a bunch of runtime stuff when not playing the game
in a v1 scene

There's a chance this all gets refactored into copy pasted v1 scripts
instead, or any v1 compatiblity gets removed entirely

For now it's useful to be able to test both player's to try replicating
player v1 movement when v2 movement is implemented very differently
(using CharacterController instead of Rigidbody). Which is possible
using this solution.