fix[Player/V2/Controller + AudioManager]: Attacks work once + Audio manager is a singleton
Attack scripts were present in two locations and they did 0 damage
because collision was handled by really old code simultaneously, so
I removed the old code and restored the damage values. I think if I
didn't do this, it would be incredibly confusing to the next person who
would implement a new physical attack.

AudioManager is null when game scene is reloaded so I made it a singleton
(the instance is static and will never be null)

I thought I committed the attack fix earlier, but I forgot and now this
commit is one which should actually be split into two commits
(fixing attacks and fixing the audio manager bug), my bad.